tools/notify-service-user-sync/requirements.txt
black==23.3.0
coverage==7.2.5
cryptography==41.0.1
python-dotenv==1.0.0
pylint==2.17.4
pytest==7.3.2
pytest-env==0.8.1
simple_salesforce==1.12.4